requirements.txt
langchain
faiss-cpu
openai
pdfplumber
python-docx
python-dotenv
streamlit
langchain-community
langchain-openai